requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
numpy==1.26.4
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
diffusers==0.31.0
transformers==4.46.0
accelerate==1.1.1
controlnet-aux==0.0.9
scikit-image==0.24.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
